反水计算修改

+++ src/main/java/com/yingliguoji/fee/po/MemberPo.java
@@ -2,12 +2,14 @@ package com.yingliguoji.fee.po;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class MemberPo {
     private Integer id;
     private String name;
-    private Integer money;
+    private BigDecimal money;
     private Integer status;
-    private Integer fs_money;
+    private BigDecimal fs_money;
     private Integer top_id;
 }
